feat(semantic-router): add support for dynamic routes in Route object
Introduced two types of routes in the `Route` object within semantic-router: static and dynamic routes. Static routes return a `Route.name` when chosen, while dynamic routes use an LLM call to produce parameter input values.

For instance, a static route might identify a query about mathematics by returning the route name, such as "math." In contrast, a dynamic route not only identifies the query but also extracts key information from the input utterance to be used in a function associated with that route.